got rid of DialoguePanel and the related tests and their instances. Making GamePanel the only panel.
