+++ src/main/java/values/TunableParameters.java
@@ -1,15 +1,12 @@
 package values;
 
-import java.awt.*;
-
 public final class TunableParameters {
 	private TunableParameters() {
 	}
 
 	public static final int SCREEN_WIDTH = 900;
 	public static final int SCREEN_HEIGHT = 600;
-	public static final int DIALOGUE_PANEL_HEIGHT = 50;
-	public static final int DIALOGUE_PANEL_WIDTH = 300;
+
 	public static final int TARGET_FPS = 45;
 	public static final int PLAYER_SPEED = 1;
 	public static final int BUTTON_WIDTH = 80;
@@ -21,7 +18,6 @@ public final class TunableParameters {
 	public static final String XML_NAME_SUFFIX = ".xml";
 
 	public static final String FILE_NAME_SUFFIX = ".txt";
-
-	public static final Color DIALOGUE_PANEL_COLOR = Color.YELLOW;
+	
 
 }
