+++ src/test/java/ui/GameFrameUITest.java
@@ -1,17 +1,15 @@
 package ui;
 
+import org.junit.Test;
+import org.mockito.Mockito;
+import values.TunableParameters;
+
+import java.awt.*;
+import java.awt.event.WindowListener;
+
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.arrayContaining;
 import static org.hamcrest.Matchers.equalTo;
-
-import java.awt.Component;
-import java.awt.Dimension;
-import java.awt.event.WindowListener;
-
-import org.junit.Test;
-import org.mockito.Mockito;
-
-import values.TunableParameters;
 
 public class GameFrameUITest {
 
@@ -23,6 +21,7 @@ public class GameFrameUITest {
 		int width = TunableParameters.SCREEN_WIDTH;
 		int height = TunableParameters.SCREEN_HEIGHT;
 		final GamePanel gamePanel = Mockito.mock(GamePanel.class);
+
 		WindowAdapterSystemExit windowAdapter = Mockito.mock(WindowAdapterSystemExit.class);
 		gameFrame = new GameFrame(gamePanel, windowAdapter) {
 			@Override
@@ -30,7 +29,7 @@ public class GameFrameUITest {
 				assertThat((GamePanel) comp, equalTo(gamePanel));
 				return gamePanel;
 			}
-
+			
 			@Override
 			public void setVisible(boolean b) {
 				assertThat(b, equalTo(true));
